Instantiated var "number" in case it was unset
In the rush to finish the formatting before the due date, I created a last-minute variable to hold the validated and formatted phone number, but I forgot to instantiate it in the case where the "phone" variable was unset. This was fixed by initializing "number" to an empty string on line 49.